requirements.txt
accelerate==0.19.0.dev0
backports.functools-lru-cache==1.6.4
cmake==3.26.3
diffusers==0.16.0.dev0
imageio==2.28.1
imageio-ffmpeg==0.4.9
jupyter==1.0.0
lit==16.0.3
lpips==0.1.4
matplotlib==3.5.3
opencv-python==4.7.0.72
pynvml==11.5.0
pyqt5-sip==12.11.0
sip==6.6.2
testpath==0.6.0
tinycudann==1.7
torch-tb-profiler==0.4.1
torchmetrics==0.11.4
transformers==4.29.0.dev0
xformers==0.0.16
